GITADORA Info Server 4.5.190415

+++ kotlin/com/prunusnira/gitadorainfo/model/SkillTable.kt
@@ -25,7 +25,7 @@ class SkillTable(var musicid: Int,
 				 var version: Int,
 				 var combo: Int,
 				 var playtime: Int,
-				 var level: Long,
+				 var level: Int,
 				 var checkfc: String,
 				 var meter: String) {
 }
